% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{delete_accelerator}
\alias{delete_accelerator}
\title{Delete an accelerator}
\usage{
delete_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of an accelerator.}
}
\description{
Delete an accelerator. Note: before you can delete an accelerator, you must disable it and remove all dependent resources (listeners and endpoint groups).
}
\section{Accepted Parameters}{
\preformatted{delete_accelerator(
  AcceleratorArn = "string"
)
}
}

