% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{delete_endpoint_group}
\alias{delete_endpoint_group}
\title{Delete an endpoint group from a listener}
\usage{
delete_endpoint_group(
  EndpointGroupArn = "string"
)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to delete.}
}
\description{
Delete an endpoint group from a listener.
}
\examples{

}
