% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{delete_listener}
\alias{delete_listener}
\title{Delete a listener from an accelerator}
\usage{
delete_listener(
  ListenerArn = "string"
)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener.}
}
\description{
Delete a listener from an accelerator.
}
\examples{

}
