% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{describe_accelerator}
\alias{describe_accelerator}
\title{Describe an accelerator}
\usage{
describe_accelerator(AcceleratorArn)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator to describe.}
}
\description{
Describe an accelerator. To see an AWS CLI example of describing an accelerator, scroll down to \strong{Example}.
}
\section{Accepted Parameters}{
\preformatted{describe_accelerator(
  AcceleratorArn = "string"
)
}
}

