% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{describe_accelerator_attributes}
\alias{describe_accelerator_attributes}
\title{Describe the attributes of an accelerator}
\usage{
describe_accelerator_attributes(AcceleratorArn = NULL)
}
\arguments{
\item{AcceleratorArn}{The Amazon Resource Name (ARN) of the accelerator with the attributes that you want to describe. Value is required.}
}
\description{
Describe the attributes of an accelerator.
}
\section{Accepted Parameters}{
\preformatted{describe_accelerator_attributes(
  AcceleratorArn = "string"
)
}
}

