% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{describe_endpoint_group}
\alias{describe_endpoint_group}
\title{Describe an endpoint group}
\usage{
describe_endpoint_group(
  EndpointGroupArn = "string"
)
}
\arguments{
\item{EndpointGroupArn}{[required] The Amazon Resource Name (ARN) of the endpoint group to describe.}
}
\description{
Describe an endpoint group.
}
\examples{

}
