% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{describe_listener}
\alias{describe_listener}
\title{Describe a listener}
\usage{
describe_listener(ListenerArn)
}
\arguments{
\item{ListenerArn}{[required] The Amazon Resource Name (ARN) of the listener to describe.}
}
\description{
Describe a listener.
}
\section{Accepted Parameters}{
\preformatted{describe_listener(
  ListenerArn = "string"
)
}
}

