% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{list_accelerators}
\alias{list_accelerators}
\title{List the accelerators for an AWS account}
\usage{
list_accelerators(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The number of Global Accelerator objects that you want to return with this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a previous call.}
}
\description{
List the accelerators for an AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_accelerators(
  MaxResults = 123,
  NextToken = "string"
)
}
}

