% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_operations.R
\name{list_listeners}
\alias{list_listeners}
\title{List the listeners for an accelerator}
\usage{
list_listeners(AcceleratorArn, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AcceleratorArn}{[required] The Amazon Resource Name (ARN) of the accelerator for which you want to list listener objects.}

\item{MaxResults}{The number of listener objects that you want to return with this call. The default value is 10.}

\item{NextToken}{The token for the next set of results. You receive this token from a previous call.}
}
\description{
List the listeners for an accelerator.
}
\section{Accepted Parameters}{
\preformatted{list_listeners(
  AcceleratorArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

