% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.globalaccelerator_package.R
\docType{package}
\name{paws.globalaccelerator-package}
\alias{paws.globalaccelerator}
\alias{paws.globalaccelerator-package}
\title{paws.globalaccelerator: AWS Global Accelerator}
\description{
This is the *AWS Global Accelerator API Reference*. This
guide is for developers who need detailed information about AWS Global
Accelerator API actions, data types, and errors. For more information
about Global Accelerator features, see the [AWS Global Accelerator
Developer
Guide](https://docs.aws.amazon.com/global-accelerator/latest/dg/Welcome.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
